feat: implemented attributes on the parser
